HTBHF-2175 move type checks to common predicates directory
