+++ src/common/predicates.js
@@ -1,4 +1,4 @@
-const { compose, isNil, isEmpty, or, not } = require('ramda')
+const { compose, isNil, isEmpty, or, not, equals, type } = require('ramda')
 
 const isNilOrEmpty = (string) => or(isNil(string), isEmpty(string))
 
@@ -6,8 +6,19 @@ const notIsNilOrEmpty = compose(not, isNilOrEmpty)
 
 const notIsNil = compose(not, isNil)
 
+const isType = t => compose(equals(t), type)
+
+const isUndefined = isType('Undefined')
+
+const isBoolean = isType('Boolean')
+
+const isString = isType('String')
+
 module.exports = {
   isNilOrEmpty,
   notIsNilOrEmpty,
-  notIsNil
+  notIsNil,
+  isUndefined,
+  isBoolean,
+  isString
 }
